% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxjitter.R
\docType{data}
\name{GeomBoxJitter}
\alias{GeomBoxJitter}
\title{A hybrid boxplot. H
Half boxplot, half scatterplot with jitter.}
\format{An object of class \code{GeomBoxJitter} (inherits from \code{GeomBoxplot}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
GeomBoxJitter
}
\arguments{
\item{jitter.colour}{Default aesthetics for jitter, set to `NULL` to inherit from the aesthetics used for the box.}

\item{jitter.width}{Width passed to `position_jitter`. Defaults to half the width of the boxplot.}

\item{jitter.height}{Height passed to `position_jitter`. Defaults to 40% of the resolution.}

\item{jitter.seed}{Seed passed to `position_jitter` for reproducible jittering.}

\item{errorbar.draw}{Draw horizontal whiskers at the top and bottom (the IQR). Defaults to `FALSE`.}

\item{errorbar.length}{Length of the horizontal whiskers (errorbar). Defaults to half the width of the half-boxplot.}
}
\description{
A hybrid boxplot. H
Half boxplot, half scatterplot with jitter.
}
\keyword{datasets}
